Add files via upload
🌍 Extinct Animals In The Earth – Group Web Assignment
This is a web-based project developed as a group assignment during our 1st Year, 2nd Semester. The project aims to raise awareness about extinct animals and their impact on the planet.

The website was created using HTML, CSS, and JavaScript and showcases information about various extinct species in an interactive and visually appealing format. This assignment helped us understand the fundamentals of web design, frontend development, and user interface creation.

As a frontend developer in this group project, I was responsible for designing and building the visual layout and interaction of the website.